Option classes in their own file. Xml doc
