+++ src/Messaging.Kafka/KafkaOptions.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using Confluent.Kafka;
-
-namespace Messaging.Kafka
+﻿namespace Messaging.Kafka
 {
     public class KafkaOptions
     {
@@ -12,42 +9,13 @@ namespace Messaging.Kafka
         /// <see href="https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md"/>
         /// </remarks>
         public KafkaProperties Properties { get; set; } = new KafkaProperties();
-
+        /// <summary>
+        /// Subscriber options
+        /// </summary>
         public SubscriberOptions Subscriber { get; set; } = new SubscriberOptions();
+        /// <summary>
+        /// Publisher options
+        /// </summary>
         public PublisherOptions Publisher { get; set; } = new PublisherOptions();
-
-        public KafkaOptions()
-        {
-
-        }
     }
-
-    public class SubscriberOptions
-    {
-
-        /// <summary>
-        /// The maximum time to block.
-        /// You should typically use a relatively short timout period because this operation cannot be cancelled
-        /// </summary>
-        public TimeSpan ConsumeTimeout { get; set; } = TimeSpan.FromSeconds(3);
-        /// <summary>
-        /// Invoked when a Kafka Consumer is instantiated
-        /// </summary>
-        public Action<Consumer<Null, byte[]>> ConsumerCreatedCallback { get; set; }
-    }
-
-    public class PublisherOptions
-    {
-
-        /// <summary>
-        /// The maximum length of time to block.
-        /// You should typically use a relatively short timout period because this operation cannot be cancelled.
-        /// </summary>
-        public TimeSpan FlushTimeout { get; set; } = TimeSpan.FromSeconds(3);
-        /// <summary>
-        /// Invoked when a Kafka Producer is instantiated
-        /// </summary>
-        public Action<Producer<Null, byte[]>> ProducerCreatedCallback { get; set; }
-    }
-
 }
