+++ src/Messaging/IMessageHandlerInfoProvider.cs
@@ -4,6 +4,9 @@ using System.Reflection;
 
 namespace Messaging
 {
+    /// <summary>
+    /// Provider of <see cref="IMessageHandler"/> reflection data.
+    /// </summary>
     public interface IMessageHandlerInfoProvider
     {
         IEnumerable<(Type messageType, Type handlerType, MethodInfo handleMethod)> GetHandlerInfo();
